Move db/model to models/index and update controllers.
