+++ seeds/clients.js
@@ -1,4 +1,4 @@
-var db = require('../db/model');
+var db = require('../models/index');
 
 db.Client
   .create({
