+++ seeds/users.js
@@ -1,4 +1,4 @@
-var db = require('../db/model');
+var db = require('../models/index');
 
 db.User
   .create({
